venus/pyproject.toml
[project]
name = "venus"
version = "0.1.0"
description = ""
authors = [{ name = "scidsgn" }]
requires-python = ">=3.12,<4"
readme = "README.md"
dependencies = [
    "pillow>=11.3.0,<12",
    "python-dotenv>=1.1.1,<2",
    "pyyaml>=6.0.2,<7",
    "music-tag>=0.4.3,<0.5",
    "fastapi>=0.117.1,<0.118",
    "uvicorn>=0.37.0,<0.38",
    "ruff>=0.13.2,<0.14",
    "websockets>=15.0.1,<16",
    "pydantic>=2.11.7,<3",
    "xmltodict>=1.0.2,<2",
    "musicbrainzngs>=0.7.1,<0.8",
    "rq>=2.6.0,<3",
    "matplotlib>=3.10.6,<4",
    "numpy>=2.3.3,<3",
    "sqlalchemy>=2.0.43,<3",
    "psycopg[binary]>=3.2.10,<4",
    "pgvector>=0.4.1,<0.5",
    "colorgram-py>=1.2.0,<2",
    "python-multipart>=0.0.20,<0.0.21",
    "pyjwt[crypto]>=2.10.1,<3",
    "librosa>=0.11.0,<0.12",
    "scipy>=1.16.2,<2",
]

[project.scripts]
start = "src.main:main"
ttml_test = "src.features.lyrics.ttml_test:main"
mb_test = "src.features.musicbrainz.test:main"
tf_test = "src.tf_test:main"
db_test = "src.db_test:main"

[tool.hatch.build.targets.sdist]
include = ["src"]

[tool.hatch.build.targets.wheel]
include = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
